refactor: 사이트 title 메타데이터 수정
