+++ src/app/layout.tsx
@@ -11,8 +11,11 @@ import '@mantine/core/styles.css'
 import './assets/style/globals.css'
 
 export const metadata: Metadata = {
-  title: 'ITMOJI - IT인들끼리 MO여 JI식을 나누는 모임 🧑🏻‍💻',
-  description: 'IT인들끼리 모여 지식을 나누는 모임 ITMOJI의 공식 홈페이지입니다.',
+  title: {
+    template: '%s | IT-MOJI',
+    default: '🧑🏻‍💻 IT인들끼리 MO여 JI식을 나누는 모임 | IT-MOJI',
+  },
+  description: 'IT인들끼리 모여 지식을 나누는 모임 IT-MOJI의 공식 홈페이지입니다.',
 }
 
 if (
